Maintain: Add panic DEBUG_CHECK list and list unit test
[panic] Added a unified debug detection interface
    -- Only the most basic part has been completed, and
        the switch between bug and warn needs to be completed.

[list] Rearranged the names of most traversal iterators
    -- Added some special linked list iteration methods.

[selftest-list] Added list unit test
    -- Build a unit test tool and pass the test.
    -- But only some basic unit tests have been done so far.

+++ kernel/sched/sched.c
@@ -18,7 +18,7 @@ struct sched_type *default_sched;
 LIST_HEAD(sched_list);
 DEFINE_PERCPU(struct sched_queue, sched_queues);
 
-static int sched_sort(struct list_head *a, struct list_head *b, void *data)
+static long sched_sort(struct list_head *a, struct list_head *b, void *data)
 {
     struct sched_type *sched_a = container_of(a, struct sched_type, list);
     struct sched_type *sched_b = container_of(b, struct sched_type, list);
